requirements-dev.txt
-r requirements.txt

# Development tools
django-extensions==3.2.3
django-debug-toolbar==4.1.0
black==23.12.0
flake8==6.1.0
isort==5.13.2
pylint==3.0.3

# Testing
pytest==7.4.3
pytest-django==4.6.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
